issue #69: inject eventTag into the event handler :tada:
