+++ core/lagompb-readside/src/main/scala/io/superflat/lagompb/readside/ReadSideProcessor.scala
@@ -2,11 +2,26 @@ package io.superflat.lagompb.readside
 
 import akka.Done
 import akka.actor.typed.ActorSystem
+import akka.cluster.sharding.typed.ShardedDaemonProcessSettings
+import akka.cluster.sharding.typed.scaladsl.ShardedDaemonProcess
+import akka.persistence.jdbc.query.scaladsl.JdbcReadJournal
+import akka.persistence.query.Offset
+import akka.projection.{ProjectionBehavior, ProjectionId}
+import akka.projection.eventsourced.scaladsl.EventSourcedProvider
+import akka.projection.eventsourced.EventEnvelope
+import akka.projection.scaladsl.{ExactlyOnceProjection, SourceProvider}
+import akka.projection.slick.SlickProjection
+import com.github.ghik.silencer.silent
 import com.google.protobuf.any
 import io.superflat.lagompb.encryption.ProtoEncryption
 import io.superflat.lagompb.protobuf.core.MetaData
+import io.superflat.lagompb.ConfigReader
+import io.superflat.lagompb.protobuf.encryption.EncryptedProto
+import org.slf4j.{Logger, LoggerFactory}
 import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
+import slick.basic.DatabaseConfig
 import slick.dbio.{DBIO, DBIOAction}
+import slick.jdbc.PostgresProfile
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success, Try}
@@ -26,23 +41,35 @@ import scala.util.{Failure, Success, Try}
  * @param ec          the execution context
  * @tparam TState the aggregate state type
  */
-abstract class ReadSideProcessor[TState <: scalapb.GeneratedMessage](encryption: ProtoEncryption)(implicit
+@silent abstract class ReadSideProcessor[TState <: scalapb.GeneratedMessage](encryption: ProtoEncryption)(implicit
     ec: ExecutionContext,
     actorSystem: ActorSystem[_]
-) extends PostgresOffsetHandler[TState](encryption) {
+) extends EventProcessor {
 
-  final override def handleEvent(
+  final val log: Logger = LoggerFactory.getLogger(getClass)
+
+  // The implementation class needs to set the akka.projection.slick config for the offset database
+  protected val dbConfig: DatabaseConfig[PostgresProfile] =
+    DatabaseConfig.forConfig("akka.projection.slick", actorSystem.settings.config)
+
+  protected val baseTag: String = ConfigReader.eventsConfig.tagName
+
+  final override def process(
       comp: GeneratedMessageCompanion[_ <: GeneratedMessage],
       event: any.Any,
+      eventTag: String,
       resultingState: any.Any,
       meta: MetaData
   ): DBIO[Done] = {
     Try {
       handle(
-        event.unpack(comp),
-        resultingState
-          .unpack[TState](aggregateStateCompanion),
-        meta
+        ReadSideEvent[TState](
+          event.unpack(comp),
+          eventTag,
+          resultingState
+            .unpack[TState](aggregateStateCompanion),
+          meta
+        )
       )
     } match {
       case Failure(exception) =>
@@ -53,10 +80,63 @@ abstract class ReadSideProcessor[TState <: scalapb.GeneratedMessage](encryption:
   }
 
   /**
+   * Initialize the projection to start fetching the events that are emitted
+   */
+  def init(): Unit = {
+    ShardedDaemonProcess(actorSystem).init[ProjectionBehavior.Command](
+      name = projectionName,
+      numberOfInstances = ConfigReader.allEventTags.size,
+      behaviorFactory = n => ProjectionBehavior(setExactlyOnceProjection(s"$baseTag$n")),
+      settings = ShardedDaemonProcessSettings(actorSystem),
+      stopMessage = Some(ProjectionBehavior.Stop)
+    )
+  }
+
+  /**
+   * Build the projection instance based upon the event tag
+   *
+   * @param tagName the event tag
+   * @return the projection instance
+   */
+  protected def setExactlyOnceProjection(
+      tagName: String
+  ): ExactlyOnceProjection[Offset, EventEnvelope[EncryptedProto]] =
+    SlickProjection
+      .exactlyOnce(
+        projectionId = ProjectionId(projectionName, tagName),
+        setSourceProvider(tagName),
+        dbConfig,
+        handler = () => new EventsReader(tagName, encryption, this)
+      )
+
+  /**
+   * Set the Event Sourced Provider per tag
+   *
+   * @param tag the event tag
+   * @return the event sourced provider
+   */
+  protected def setSourceProvider(tag: String): SourceProvider[Offset, EventEnvelope[EncryptedProto]] =
+    EventSourcedProvider
+      .eventsByTag[EncryptedProto](actorSystem, readJournalPluginId = JdbcReadJournal.Identifier, tag)
+
+  /**
+   * The projection Name must be unique
+   *
+   * @return
+   */
+  def projectionName: String
+
+  /**
+   * aggregate state. it is a generated scalapb message extending the LagompbState trait
+   *
+   * @return aggregate state
+   */
+  def aggregateStateCompanion: scalapb.GeneratedMessageCompanion[TState]
+
+  /**
    * Handles aggregate event persisted and made available for read model
    *
    * @param event the aggregate event
-   * @param state the Lagompb state that wraps the actual state and some meta data
    */
-  def handle(event: scalapb.GeneratedMessage, state: TState, metaData: MetaData): DBIO[Done]
+  def handle(event: ReadSideEvent[TState]): DBIO[Done]
 }
